# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ReplitReasoner API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface AgentRunInput {
  /**
   * The programming task for the agent to solve
   * @minLength 1
   */
  task: string;
}

export type AgentRunStatus =
  (typeof AgentRunStatus)[keyof typeof AgentRunStatus];

export const AgentRunStatus = {
  running: "running",
  completed: "completed",
  failed: "failed",
} as const;

export interface AgentRun {
  id: number;
  task: string;
  status: AgentRunStatus;
  createdAt: string;
  stepCount: number;
}

export type AgentRunDetailStatus =
  (typeof AgentRunDetailStatus)[keyof typeof AgentRunDetailStatus];

export const AgentRunDetailStatus = {
  running: "running",
  completed: "completed",
  failed: "failed",
} as const;

export type AgentStepType = (typeof AgentStepType)[keyof typeof AgentStepType];

export const AgentStepType = {
  thought: "thought",
  action: "action",
  observation: "observation",
  complete: "complete",
  error: "error",
} as const;

export interface AgentStep {
  id: number;
  runId: number;
  type: AgentStepType;
  content: string;
  cycle: number;
  createdAt: string;
}

export interface AgentRunDetail {
  id: number;
  task: string;
  status: AgentRunDetailStatus;
  createdAt: string;
  steps: AgentStep[];
}

export interface RunStats {
  totalRuns: number;
  completedRuns: number;
  failedRuns: number;
  runningRuns: number;
  totalSteps: number;
}