# lib/api-zod/src/generated/types/agentRun.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ReplitReasoner API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AgentRunStatus } from "./agentRunStatus";

export interface AgentRun {
  id: number;
  task: string;
  status: AgentRunStatus;
  createdAt: string;
  stepCount: number;
}