# lib/api-zod/src/generated/types/agentRunDetail.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ReplitReasoner API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AgentRunDetailStatus } from "./agentRunDetailStatus";
import type { AgentStep } from "./agentStep";

export interface AgentRunDetail {
  id: number;
  task: string;
  status: AgentRunDetailStatus;
  createdAt: string;
  steps: AgentStep[];
}