# lib/api-zod/src/generated/types/agentRunInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ReplitReasoner API specification
 * OpenAPI spec version: 0.1.0
 */

export interface AgentRunInput {
  /**
   * The programming task for the agent to solve
   * @minLength 1
   */
  task: string;
}