# lib/api-zod/src/generated/types/agentStep.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ReplitReasoner API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AgentStepType } from "./agentStepType";

export interface AgentStep {
  id: number;
  runId: number;
  type: AgentStepType;
  content: string;
  cycle: number;
  createdAt: string;
}