# lib/api-zod/src/generated/types/runStats.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ReplitReasoner API specification
 * OpenAPI spec version: 0.1.0
 */

export interface RunStats {
  totalRuns: number;
  completedRuns: number;
  failedRuns: number;
  runningRuns: number;
  totalSteps: number;
}